fpathconf(2) returns long.

+++ lib/libc_r/uthread/uthread_fpathconf.c
@@ -1,7 +1,7 @@
 /*
  * David Leonard <d@openbsd.org>, 1999. Public Domain.
  *
- * $OpenBSD: uthread_fpathconf.c,v 1.1 1999/01/08 05:42:18 d Exp $
+ * $OpenBSD: uthread_fpathconf.c,v 1.2 2002/01/02 16:19:35 fgsch Exp $
  */
 
 #include <unistd.h>
@@ -12,7 +12,7 @@
 long
 fpathconf(int fd, int name)
 {
-	int             ret;
+	long	ret;
 
 	if ((ret = _FD_LOCK(fd, FD_READ, NULL)) == 0) {
 		ret = _thread_sys_fpathconf(fd, name);
